feat: Update next-intl-shadcn-form package and index.d.ts
This commit updates the next-intl-shadcn-form package to version 0.0.4 and modifies the index.d.ts file. The changes in index.d.ts include importing ReactNodeArray from react and updating the NextIntlTranslationFunction type to include the rich and markup methods. These changes enhance the flexibility and functionality of the translation function in next-intl.